+++ precompiles/staking/tx_test.go
@@ -51,7 +51,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				return []interface{}{}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 6, 0),
 		},
@@ -69,7 +69,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"is not the same as delegator address",
 		},
@@ -86,7 +86,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid description",
 		},
@@ -103,7 +103,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid commission",
 		},
@@ -120,7 +120,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid amount",
 		},
@@ -137,7 +137,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid validator address",
 		},
@@ -154,7 +154,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid type for",
 		},
@@ -171,7 +171,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"illegal base64 data",
 		},
@@ -188,7 +188,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"consensus pubkey len is invalid",
 		},
@@ -205,7 +205,7 @@ func (s *PrecompileTestSuite) TestCreateValidator() {
 				}
 			},
 			200000,
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"invalid amount",
 		},
@@ -311,12 +311,12 @@ func (s *PrecompileTestSuite) TestDelegate() {
 	}{
 		{
 			"fail - empty input args",
-			func(operatorAddress string) []interface{} {
+			func(string) []interface{} {
 				return []interface{}{}
 			},
 			200000,
 			big.NewInt(0),
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 3, 0),
 		},
@@ -346,7 +346,7 @@ func (s *PrecompileTestSuite) TestDelegate() {
 			},
 			200000,
 			big.NewInt(1),
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			fmt.Sprintf(cmn.ErrInvalidDelegator, ""),
 		},
@@ -361,7 +361,7 @@ func (s *PrecompileTestSuite) TestDelegate() {
 			},
 			200000,
 			big.NewInt(1),
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			fmt.Sprintf(cmn.ErrInvalidAmount, nil),
 		},
@@ -378,7 +378,7 @@ func (s *PrecompileTestSuite) TestDelegate() {
 			},
 			200000,
 			big.NewInt(15),
-			func(data []byte) {},
+			func([]byte) {},
 			true,
 			"insufficient funds",
 		},
@@ -401,7 +401,7 @@ func (s *PrecompileTestSuite) TestDelegate() {
 		//	},
 		//	200000,
 		//	big.NewInt(15),
-		//	func(data []byte) {},
+		//	func( []byte) {},
 		//	true,
 		//	"cannot delegate/undelegate",
 		// },
@@ -521,10 +521,10 @@ func (s *PrecompileTestSuite) TestUndelegate() {
 	}{
 		{
 			"fail - empty input args",
-			func(operatorAddress string) []interface{} {
+			func(string) []interface{} {
 				return []interface{}{}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(0),
 			true,
@@ -554,7 +554,7 @@ func (s *PrecompileTestSuite) TestUndelegate() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -569,7 +569,7 @@ func (s *PrecompileTestSuite) TestUndelegate() {
 					nil,
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -650,10 +650,10 @@ func (s *PrecompileTestSuite) TestRedelegate() {
 	}{
 		{
 			"fail - empty input args",
-			func(srcOperatorAddr, dstOperatorAddr string) []interface{} {
+			func(string, string) []interface{} {
 				return []interface{}{}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(0),
 			true,
@@ -685,7 +685,7 @@ func (s *PrecompileTestSuite) TestRedelegate() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -701,7 +701,7 @@ func (s *PrecompileTestSuite) TestRedelegate() {
 					nil,
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -717,7 +717,7 @@ func (s *PrecompileTestSuite) TestRedelegate() {
 					big.NewInt(-1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -797,10 +797,10 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 	}{
 		{
 			"fail - empty input args",
-			func(operatorAddress string) []interface{} {
+			func(string) []interface{} {
 				return []interface{}{}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(0),
 			true,
@@ -816,7 +816,7 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -832,7 +832,7 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 					nil,
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -848,7 +848,7 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -864,7 +864,7 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,
@@ -880,7 +880,7 @@ func (s *PrecompileTestSuite) TestCancelUnbondingDelegation() {
 					big.NewInt(1),
 				}
 			},
-			func(data []byte) {},
+			func([]byte) {},
 			200000,
 			big.NewInt(1),
 			true,